almost complete endpoints

+++ src/controllers/categories.ts
@@ -2,7 +2,8 @@ import { asyncHandler } from '../utils/asyncHandler.js'
 import { ApiResponse } from '../utils/ApiResponse.js'
 import { ApiError } from '../utils/ApiError.js'
 import { Product } from "../models/product.models.js";
-import { Category } from  "../models/categories.models.js"   // type fix leter spelling 
+import { Category } from  "../models/categories.models.js" 
+import { isValidObjectId } from 'mongoose';
 import { nodeCache } from '../app.js';
 
 
@@ -63,6 +64,9 @@ const editCategory = asyncHandler( async(req,res) => {
  // Todo: edite category oonly admin 
 
     const categoryId = req.params
+
+    if(! isValidObjectId(categoryId)) throw new ApiError(400,"invalid category Id");
+     
     const name = req.body
     const category = await Category.findByIdAndUpdate({_id:categoryId},{
         $set:{
@@ -82,20 +86,41 @@ const deleteCategory = asyncHandler( async(req,res) => {
  // Todo: delete category oonly admin 
   const categoryId = req.params
 
-  const deleteCategory = await Category.findByIdAndDelete(categoryId)
+  const isCategoryExisted = await Category.findById(categoryId);
 
- return res
- .status(200)
- .json(
-    new ApiResponse(200,{}," categories is deleted successfully") 
- )
+  if(!isCategoryExisted) throw new ApiError(404,"category does't exists")
+
+    await Category.findByIdAndDelete(categoryId)
+
+        return res
+        .status(200)
+        .json(
+            new ApiResponse(200,{}," categories is deleted successfully") 
+        )
 })
 
+const searchCategory = asyncHandler( async(req,res) => {
 
+
+    const { category } = req.query
+
+    
+    const searchCategory = await Category.find({ name:{ $regex:category , $options:"i"}  })
+
+    if(!searchCategory) throw new ApiError(404," there is not category exist in This site")
+
+
+    return res
+    .status(200)
+    .json(
+       new ApiResponse(200,searchCategory," categories is deleted successfully") 
+    )
+})
 
 export {
     getAllCategories,
     addCategory,
     editCategory,
-    deleteCategory
+    deleteCategory,
+    searchCategory
 }
